multipart,putmapping

+++ job/src/main/java/com/kiruthika/job/Service/JobListingService.java
@@ -7,10 +7,8 @@ import org.springframework.stereotype.Service;
 
 import com.kiruthika.job.Entity.Employer;
 import com.kiruthika.job.Entity.JobList;
-import com.kiruthika.job.Entity.UserData;
 import com.kiruthika.job.Repository.EmployerRepository;
 import com.kiruthika.job.Repository.JobListRepository;
-import com.kiruthika.job.Repository.UserDataRepository;
 
 @Service
 public class JobListingService {
@@ -25,8 +23,8 @@ public class JobListingService {
         return jobListRepository.findAll();
     }
 
-    public JobList postJobs(JobList jobList) {
-        return jobListRepository.save(jobList);
+    public JobList postJobs(JobList jobList)throws Exception {
+            return jobListRepository.save(jobList);
     }
 
     public List<JobList> getJobs(String uname) {
@@ -38,12 +36,36 @@ public class JobListingService {
         return jobListRepository.findByTitle(title);
     }
 
+    public List<JobList> getJobsByCategory(String category){
+        return jobListRepository.findByCategory(category);
+    }
+
     public void deleteJob(Long jobId) {
-        jobListRepository.deleteById(jobId);
+        if (jobListRepository.existsById(jobId)) {
+            jobListRepository.deleteById(jobId);
+        } 
+       else {
+            throw new RuntimeException("Job with ID " + jobId + " not found");
+        }
     }
 
     public void deleteAllJobs(){
         jobListRepository.deleteAll();
     }
-        
+      
+    public JobList editJob(Long jobId, JobList updatedJob) {
+        JobList existingJob = jobListRepository.findById(jobId).get();
+
+        if (existingJob != null) {
+            existingJob.setTitle(updatedJob.getTitle());
+            existingJob.setDescription(updatedJob.getDescription());
+            existingJob.setRequirements(updatedJob.getRequirements());
+            existingJob.setCategory(updatedJob.getCategory());
+            existingJob.setDeadline(updatedJob.getDeadline());
+            existingJob.setLocation(updatedJob.getLocation());
+            return jobListRepository.save(existingJob);
+        } else {
+            return null; 
+        }
+    }
 }
